Fixed EVERYTHING
Wow it's been a while huh. Well, everything was broken and I fixed it. The last time I updated this code, I was in high school and looking back on this 5 years later, I realize how absolutely garbage I was at coding. Anyways, the main changes were a shift from web scraping to solely using API calls, colorful debugging logs, creation of a config file, and the use of a template presentation that allows the design to be easily changed.